+++ packages/cosmwasm/src/cosmwasmclient.spec.ts
@@ -3,7 +3,7 @@ import { Sha256 } from "@cosmjs/crypto";
 import { Bech32, fromHex, fromUtf8, toAscii, toBase64 } from "@cosmjs/encoding";
 import {
   assertIsBroadcastTxSuccess,
-  makeSignBytes,
+  makeStdSignDoc,
   MsgSend,
   Secp256k1Wallet,
   StdFee,
@@ -236,8 +236,8 @@ describe("CosmWasmClient", () => {
 
       const chainId = await client.getChainId();
       const { accountNumber, sequence } = await client.getSequence(alice.address0);
-      const signBytes = makeSignBytes([sendMsg], fee, chainId, memo, accountNumber, sequence);
-      const signature = await wallet.sign(alice.address0, signBytes);
+      const signDoc = makeStdSignDoc([sendMsg], fee, chainId, memo, accountNumber, sequence);
+      const signature = await wallet.sign(alice.address0, signDoc);
       const signedTx = {
         msg: [sendMsg],
         fee: fee,