+++ packages/launchpad/src/lcdapi/staking.spec.ts
@@ -3,7 +3,7 @@ import { assert, sleep } from "@cosmjs/utils";
 
 import { coin, coins } from "../coins";
 import { assertIsBroadcastTxSuccess } from "../cosmosclient";
-import { makeSignBytes } from "../encoding";
+import { makeStdSignDoc } from "../encoding";
 import { MsgDelegate, MsgUndelegate } from "../msgs";
 import { Secp256k1Wallet } from "../secp256k1wallet";
 import { SigningCosmosClient } from "../signingcosmosclient";
@@ -46,8 +46,8 @@ describe("StakingExtension", () => {
         };
         const memo = "Test delegation for wasmd";
         const { accountNumber, sequence } = await client.getSequence();
-        const signBytes = makeSignBytes([msg], defaultFee, chainId, memo, accountNumber, sequence);
-        const signature = await wallet.sign(faucet.address, signBytes);
+        const signDoc = makeStdSignDoc([msg], defaultFee, chainId, memo, accountNumber, sequence);
+        const signature = await wallet.sign(faucet.address, signDoc);
         const tx = {
           msg: [msg],
           fee: defaultFee,
@@ -69,8 +69,8 @@ describe("StakingExtension", () => {
         };
         const memo = "Test undelegation for wasmd";
         const { accountNumber, sequence } = await client.getSequence();
-        const signBytes = makeSignBytes([msg], defaultFee, chainId, memo, accountNumber, sequence);
-        const signature = await wallet.sign(faucet.address, signBytes);
+        const signDoc = makeStdSignDoc([msg], defaultFee, chainId, memo, accountNumber, sequence);
+        const signature = await wallet.sign(faucet.address, signDoc);
         const tx = {
           msg: [msg],
           fee: defaultFee,