+++ packages/launchpad/src/signer.ts
@@ -1,3 +1,4 @@
+import { StdSignDoc } from "./encoding";
 import { StdSignature } from "./types";
 
 export type PrehashType = "sha256" | "sha512" | null;
@@ -19,6 +20,13 @@ export interface OfflineSigner {
 
   /**
    * Request signature from whichever key corresponds to provided bech32-encoded address. Rejects if not enabled.
+   *
+   * @param signerAddress The address of the account that should sign the transaction
+   * @param signDoc The content that should be signed
    */
-  readonly sign: (address: string, message: Uint8Array, prehashType?: PrehashType) => Promise<StdSignature>;
+  readonly sign: (
+    signerAddress: string,
+    signDoc: StdSignDoc,
+    prehashType?: PrehashType,
+  ) => Promise<StdSignature>;
 }